Moved mpe_channel_command, mpe_service_metadata and contract to shared package
Refactored SDK package to use shared code
