+++ packages/sdk/snet/sdk/service_client.py
@@ -25,7 +25,7 @@ class ServiceClient:
         self.group = group
         self.metadata = metadata
         self.payment_channel_management_strategy = payment_channel_management_strategy
-        self.expiry_threshold = self.metadata.payment_expiration_threshold
+        self.expiry_threshold = self.metadata["payment_expiration_threshold"]
         self._base_grpc_channel = self._get_grpc_channel()
         self.grpc_channel = grpc.intercept_channel(self._base_grpc_channel, generic_client_interceptor.create(self._intercept_call))
         self.payment_channel_state_service_client = self._generate_payment_channel_state_service_client()
@@ -46,7 +46,7 @@ class ServiceClient:
     def _get_grpc_channel(self):
         endpoint = self.options.get("endpoint", None)
         if endpoint is None:
-            endpoint = next(filter(lambda endpoint: endpoint["group_name"] == self.group["group_name"], self.metadata["endpoints"]))["endpoint"]
+            endpoint = self.metadata.get_endpoints_for_group(self.group["group_name"])[0]
         endpoint_object = urlparse(endpoint)
         if endpoint_object.port is not None:
             channel_endpoint = endpoint_object.hostname + ":" + str(endpoint_object.port)
@@ -63,7 +63,7 @@ class ServiceClient:
 
     def _get_service_call_metadata(self):
         channel = self.payment_channel_management_strategy.select_channel(self)
-        amount = channel.state["last_signed_amount"] + int(self.metadata.pricing["price_in_cogs"])
+        amount = channel.state["last_signed_amount"] + int(self.metadata["pricing"]["price_in_cogs"])
         message = web3.Web3.soliditySha3(
             ["address", "uint256", "uint256", "uint256"],
             [self.sdk.mpe_contract.contract.address, channel.channel_id, channel.state["nonce"], amount]
