+++ packages/snet_cli/snet/snet_cli/mpe_service_metadata.py
