+++ packages/snet_cli/snet_cli/mpe_client_command.py
@@ -6,10 +6,10 @@ import sys
 import grpc
 from eth_account.messages import defunct_hash_message
 
-from snet_cli.mpe_channel_command import MPEChannelCommand
 from snet_cli.utils_agi2cogs import cogs2stragi
 
 from snet.snet_cli.utils import open_grpc_channel, rgetattr, compile_proto, RESOURCES_PATH
+from snet.snet_cli.mpe_channel_command import MPEChannelCommand
 from snet.snet_cli.utils_proto import import_protobuf_from_dir, switch_to_json_payload_encoding
